moved reflect to model.
