+++ task/task.go
@@ -14,7 +14,7 @@ import (
 	"github.com/konveyor/tackle2-hub/auth"
 	crd "github.com/konveyor/tackle2-hub/k8s/api/tackle/v1alpha1"
 	"github.com/konveyor/tackle2-hub/model"
-	"github.com/konveyor/tackle2-hub/reflect"
+	"github.com/konveyor/tackle2-hub/model/reflect"
 	"github.com/konveyor/tackle2-hub/settings"
 	"gorm.io/gorm"
 	"gorm.io/gorm/clause"
